+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added `craft\cache\FileCache`.
 - Added `craft\helpers\ArrayHelper::whereIn()`.
 - Added `craft\helpers\Cp::elementHtml()`.
+- Added `craft\models\Site::getLocale()`.
 - Added `craft\services\ProjectConfig::DIFF_CACHE_KEY`.
 - Added `craft\services\Sites::getGroupByUid()`.